+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
 
+### Changed
+
+- RandomSkunk.Results
+    - Extract more information from an `Exception` when creating an `Error` from it.
+
 ## [1.0.0-alpha21] - 2022-11-10
 
 [RandomSkunk.Results API](https://www.fuget.org/packages/RandomSkunk.Results/1.0.0-alpha21) ([diff](https://www.fuget.org/packages/RandomSkunk.Results/1.0.0-alpha21/lib/netstandard2.1/diff/1.0.0-alpha20/))